Ch.1 data out of functions: Fix typos.
Note that my text editor has changed tabs in between lines to spaces,
so the number of changes in this commit is larger than the amount that
was meaningfully changed. This may be the case for future commits as
well, but I will no longer make an indication of this happening.